feat: move seo into the resume block
