+++ src/404/index.tsx
@@ -1,7 +1,7 @@
 import { Link } from '@tanstack/react-router';
 import { cx } from '../../styled-system/css';
 import * as commonStyles from '../resume/common.styles';
-import SEO from '../seo';
+import SEO from '../resume/seo';
 
 export default function ErrorPage() {
   return (
