+++ src/resume/seo.test.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import SEO, { SEOProps } from '.';
+import SEO, { SEOProps } from './seo';
 
 describe('SEO Rendering', () => {
   it('Should match snapshot', () => {
