+++ src/resume/seo.tsx
